Moved analytics from /utils to /lib/analytics as utils seem to contain pure functions/no side effects. We can think about where to put analytics later, and whether to add to a store for easier testing similar to the Calypso analytics actions
