+++ client/landing/gutenboarding/lib/analytics/index.ts
@@ -6,7 +6,7 @@ import { recordTracksEvent } from '@automattic/calypso-analytics';
 /**
  * Internal dependencies
  */
-import { FLOW_ID } from '../constants';
+import { FLOW_ID } from '../../constants';
 
 /**
  * Make tracks call with embedded flow.
@@ -38,6 +38,8 @@ export function recordOnboardingStart( ref = '' ): void {
  * Analytics call at the completion  of a Gutenboarding flow
  *
  * @param {object} params A set of params to pass to analytics for signup completion
+ * @param {boolean} params.isNewUser Whether the user is newly signed up
+ * @param {boolean} params.isNewSite Whether a new site is created in the flow
  */
 export function recordOnboardingComplete( { isNewUser = false, isNewSite = true } ): void {
 	trackEventWithFlow( 'calypso_signup_complete', {
